requirements.txt
baumer-neoapi @ file:///H:/HK242/LVTN/Hardware/Baumer_neoAPI_1.4.1_win_x86_64_python/wheel/baumer_neoapi-1.4.1-cp34.cp35.cp36.cp37.cp38.cp39.cp310.cp311.cp312-none-win_amd64.whl
bcrypt==4.3.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.7
cycler==0.12.1
et_xmlfile==2.0.0
filelock==3.13.1
fonttools==4.59.2
fsspec==2024.6.1
idna==3.10
Jinja2==3.1.3
joblib==1.5.2
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.3
numpy==2.2.6
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pillow==11.0.0
pycparser==2.23
pymelsec==0.2.5
pyparsing==3.2.4
PyQt6==6.9.1
PyQt6-Qt6==6.9.2
PyQt6_sip==13.10.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sympy==1.13.1
threadpoolctl==3.6.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
typing_extensions==4.9.0
tzdata==2025.2
urllib3==2.5.0
